pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-yasg = "^1.21.7"
factory-boy = "^3.3.0"
djangorestframework-simplejwt = "^5.3.0"
coveralls = "^3.3.1"
rollbar = "^0.16.3"
newrelic = "^9.1.0"
django-storages = "^1.14.1"
boto3 = "^1.28.59"
pillow = "^10.0.1"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest-django = "^4.5.2"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
show_missing = true
fail_under = 80
